+++ corso01/sketch37/sketch37.ino
@@ -36,7 +36,7 @@ int valore = 0;                   // valore restituito dall'analogRead()
 float temperatura = 0;            // variabile per conservare
                                   // la temperatura rilevata
 
-int misure[8];                    // array in cui memorizzare 8 valori
+float misure[8];                    // array in cui memorizzare 8 valori
 
 void setup() {
   lcd.begin(16, 2);           //impostiamo il numero di colonne ed il numero di righe di lcd
